Add edit button to creature list
