+++ src/app/creatures/page.tsx
@@ -1,6 +1,7 @@
 import { QueryClient } from "@tanstack/react-query";
 import { getInitialCreatures } from "@/services/creatures";
-import CreatureList from "@/components/creatures/creature-list";
+import { DataTable } from "./components/data-table";
+import { creatureColumns } from "./components/columns";
 
 export default async function Creatures() {
 	const queryClient = new QueryClient();
@@ -10,9 +11,5 @@ export default async function Creatures() {
 		queryFn: getInitialCreatures,
 	});
 
-	return (
-		<div>
-			<CreatureList />
-		</div>
-	);
+	return <DataTable columns={creatureColumns} />;
 }
